Setup Profile page; Setup registration rerout
