+++ src/components/Login.jsx
@@ -58,7 +58,9 @@ const LoginPage = (props) => {
     
     // Check if the user is already logged in
     if (props.profile.isLoggedIn || done) {
-        return (<Redirect to="/dashboard" />);
+        if (props.profile.registration_status === "unregistered") {
+            return (<Redirect to="/profile" />);
+        } else return (<Redirect to="/dashboard" />);
     }
 
     return (
