+++ src/components/Pages.jsx
@@ -1,6 +1,6 @@
 import LandingPage from "./Landing/Landing";
 import DashboardPage from "./Dashboard/Dashboard";
-import ApplicationPage from "./Dashboard/Application";
+import ProfilePage from "./Dashboard/Profile";
 import LoginPage from "./Login";
 import SignUpPage from "./Signup";
 import ForgotPage from "./Forgot";
@@ -13,7 +13,7 @@ import TeamPage from "./Team/Team";
 export {
     LandingPage,
     DashboardPage,
-    ApplicationPage,
+    ProfilePage,
     LoginPage,
     SignUpPage,
     MagicPage,
